Fork network validation from node-config
Only validate things the sdn plugin will use, reduces dependency
set.
